moved api file for team page
